+++ src/lib/components/editor/CaretContext.ts
@@ -17,6 +17,60 @@ import type Node from '$lib/models/chapter/Node';
 
 export type PasteContent = { plain: string | undefined };
 
+export type Accent = '`' | "'" | '^' | '~' | ':';
+export const Accents: Record<string, string> = {
+    '`a': 'à',
+    '`A': 'À',
+    '`e': 'è',
+    '`E': 'È',
+    '`i': 'ì',
+    '`I': 'Ì',
+    '`o': 'ò',
+    '`O': 'Ò',
+    '`u': 'ù',
+    '`U': 'Ù',
+    "'a": 'á',
+    "'A": 'Á',
+    "'e": 'é',
+    "'E": 'É',
+    "'i": 'í',
+    "'I": 'Í',
+    "'o": 'ó',
+    "'O": 'Ó',
+    "'u": 'ú',
+    "'U": 'Ú',
+    '^a': 'â',
+    '^A': 'Â',
+    '^e': 'ê',
+    '^E': 'Ê',
+    '^i': 'î',
+    '^I': 'Î',
+    '^o': 'ô',
+    '^O': 'Ô',
+    '^u': 'û',
+    '^U': 'Û',
+    '~a': 'ã',
+    '~A': 'Ã',
+    '~e': 'ẽ',
+    '~E': 'Ẽ',
+    '~i': 'ĩ',
+    '~I': 'Ĩ',
+    '~o': 'õ',
+    '~O': 'Õ',
+    '~u': 'ũ',
+    '~U': 'Ũ',
+    ':a': 'ä',
+    ':A': 'Ä',
+    ':e': 'ë',
+    ':E': 'Ë',
+    ':i': 'ï',
+    ':I': 'Ï',
+    ':o': 'ö',
+    ':O': 'Ö',
+    ':u': 'ü',
+    ':U': 'Ü',
+};
+
 type CaretContext = {
     chapter: boolean;
     range: CaretRange;
@@ -50,6 +104,8 @@ type CaretContext = {
         text: PasteContent | Node,
         save: boolean,
     ) => Edit;
+    setAccent: (accent: Accent | undefined) => void;
+    getAccent: () => Accent | undefined;
 };
 
 export type { CaretContext as default };